pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lme-al-market-signal-extractor"
version = "0.1.0"
description = "Core LangGrpah utilities for LME Aluminum market signals"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }

authors = [
    { name = "Tanul Kumar Srivastava" }
]

dependencies = [
    "numpy==1.26.4",
    "pandas==2.2.2",
    "langgraph>=0.2.0,<0.3.0",
    "langchain>=0.2.0,<0.3.0",
    "python-dotenv==1.0.1",
    "langchain-community>=0.2.0,<0.3.0",
    "tqdm==4.67.3",
    "langchain-groq==0.1.9"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]